+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Draggable was reintegrated with the new Props collections for the OnPress/OnEnter/OnExit commands
 - WTreeManager get the transition commands from the InteractionHandler and is responsible for executing them
 - Text Widget now updates the space when the FontSize or the Font changes (in the setters), instead of the props
+- JustifyAround with only one child puts it at the center instead of doing nothing. It makes more sense this way since it's supposed to put empty space around widgets
 ### Removed
 - Tree collection
 - WidgetTree